requirements.txt
fastapi==0.111.0
jiwer==3.0.4
more_itertools==10.2.0
numba==0.59.1
numpy==1.26.4
python-dotenv==1.0.1
python-socketio==5.11.2
regex==2024.5.15
tiktoken==0.7.0
torch
tqdm==4.66.4
uvicorn==0.30.1
openai
pymongo
joblib
google-generativeai==0.7.2